+++ DTCD-StyleSystem/src/base-components/BaseCodeEditor/BaseCodeEditor.scss
@@ -107,6 +107,26 @@
   }
 }
 
+.CodeMirror-hover-tooltip {
+  background-color: infobackground;
+  border: 1px solid var(--title);
+  border-radius: 4px 4px 4px 4px;
+  // color: infotext;
+  // font-family: monospace;
+  // font-size: 10pt;
+  overflow: hidden;
+  padding: 2px 5px;
+  position: fixed;
+  z-index: 100;
+  max-width: 600px;
+  opacity: 0.1;
+  transition: opacity .2s;
+
+  ul {
+    padding-left: 20px;
+  }
+}
+
 // styles for CodeMirror 6
 /*
 .BaseInput {
